+++ djangocms_form_builder/static/djangocms_form_builder/js/actions_form.js
@@ -1,7 +1,12 @@
 $(function () {
+    'use strict';
     $('fieldset.action-auto-hide input[type="checkbox"][name="form_actions"]').each(function (index, element) {
+        const target = $('.' + $(element).attr("value"));
         if (element.checked) {
-            $("."+$(element).attr("value")).removeClass("action-hide");
+            target.removeClass("action-hide");
+        }
+        if (!target.find('.form-row:not(.hidden)').length) {
+            target.addClass("empty");
         }
         $(element).on("change", function (event) {
             var element = event.target;